requirements.txt
alembic==1.16.5
blinker==1.9.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.1.8
cryptography==42.0.0
Flask==2.3.3
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.0.5
Flask-SocketIO==5.6.0
Flask-WTF==1.2.2
fpdf2==2.8.3
greenlet==3.2.4
gunicorn==21.2.0
h11==0.16.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
Pillow==11.3.0
psycopg2-binary==2.9.10
pdfkit==1.0.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-engineio==4.13.0
python-socketio==5.16.0
pytz==2022.7.1
qrcode==8.2
reportlab==4.4.7
requests==2.31.0
simple-websocket==1.1.0
SQLAlchemy==2.0.45
tomli==2.3.0
typing-extensions==4.15.0
urllib3==2.6.3
Werkzeug==2.3.7
WeasyPrint==59.0
wsproto==1.2.0
WTForms==3.2.1
WTForms-SQLAlchemy==0.3
email_validator==2.1.0
psutil==5.9.8
z